AGDIGGER-73 Allow for parameterized builds from the DiggerClient
This change allows parameters to be defined when creating a Job,
allowing for parameterized builds to be performed. This is done
by providing a fourth parameter to diggerClient.createJob(), a
List of BuildParameter objects.

The BuildParameter contains the name of the parameter, the description
for the parameter and the default value for the parameter.

The builds can then be overridden when calling diggerClient.build()
by providing a third parameter, a Map containing the name of the
parameter as the key and the value of the parameter as the value.

The job.xml file has also been updated to loop through the list
of parameters if they are provided and add these to the config
file.

Currently only String type parameters are supported.